Next steps

+++ src/ViennaRNA/mfe.h
@@ -298,7 +298,7 @@ vrna_backtrack_window(vrna_fold_compound_t  *fc,
                       const char            *Lfold_filename,
                       long                  file_pos,
                       char                  **structure,
-                      float                 mfe);
+                      double                mfe);
 
 /**
  * End backtracking related interfaces
